Add setter for description, status and sequence
This will add a new roundtrip test that checks for an extended set
of properties alongside a new testfile "jscalendar_extedned.json".

Also adds new methods to parse the jmap properties "description",
"status" and "sequence" to their iCal counterparts.

Fixes a mistake where the getter for the sequence property returned
an entire vObject object instead of just the integer value associated
to it.