Add mapping multiple events both ways
This will chagne the code in the mapToJmap method to properly map
multiple iCal events (including modified occurences) to the
recurrenceOverride property of their respective master event.

It will also add code to the mapFromJmap method that allows for
recurrenceOverrides to be mapped to additional VEVENTs in the iCal
file, using properties of their master event to fill in blanks.

A new json file with recurrenceOverrides and some test cases still
need to be added for these features.