Add propper mapping of recurrenceOverrides to seperate VEVENTS
This will change the mapFromJmap code as to where it no longer maps
recurrenceOverride properties to seperate VCalendar components but adds
them to the VCalendar of their respective master event to then be added
to the iCal data array at once.

This was necessary to not lose any data regarding those events and makes
it easier to check the mapping in the roundtrip test.

Also adds a new getVevent() method to the adapter that will return the
VEVENT component as a serialized string to paste it into the VCalendar
component.