Fix issues addressed in PR
This will change the way mapFromJmap() handles recurrenceOverrides
by adding a new getVeventComponents() method to the adapter that
returns the VEVENT component of an iCal event as an associative array.
The array is then added to the master event by using the VObject
library.

This will also change the mapToJmap() method by moving the step
to set the recurrenceOverride properties of the JSCal event outside
of the for loop it was previously in.

Logging is also added to the JSCalendarICalendarAdapterUtil class
when parsing does not work.

Lastly, changed some of the comments to read more clearly.