Correct mapping of alert trigger properties
This will correct the way that the properties of an alert are mapped.

Namely, instead of simply changing the keyword of the parameter, a new VALARM
is added before converting the alert to iCal and the properties RELATED=START/END
and VALUE=DATE-TIME  are added using the correct sabre/VObject methods. This
caused some issues when testing before. Also, the RELATED property is now mapped
correctly to an OffsetTriger.

Modifed the demo data and test cases accordingly. All works as expected now.