Add setter for participants
This will add the setter for participants into the adapter code. It works
by retreiving the value(mailto/other) from the JSCal object and parsing the
parameters via newly added helper functions. The value for role is used
to distingush between ATTENDEE and ORGANIZER.

It will also add functions to the existing util class to help convert
the values between the formats.

Changed the way the getter parses the roles to now be an associative array
with the key being the role and the value a boolean(true) to conform to
the JSCalendar RFC.