Add changes to comments and participant switch
This will rewrites to a comment mentioned in the previous PR regarding
the mapping of iCal ACTION values to the JSCal action property.

Also changes a comment on how participants are created when it comes
to having an ORGANIZR also be an ATTENDEE or not.

Lastly adds a break to the extractParticipantPrameters switch case that
caused make fulltest to fail.