Fix issue regarding RRULE until value conversion
The logger used in the JSCalendarICalendarAdapterUtil methods regarding
the until parameter of recurrence Rules was not initialized properly.

Also, there was an issue regarding the creation of DateTime objects from
the iCal value as it was always read as a UTC date time. The format can now
be UTC, local with time and local without time, as its format is linked to
the events "DTSTART" property.